Add Odometries to DB, including GT

+++ src/models/test_definition.rs
@@ -29,8 +29,8 @@ pub struct TestDefinitionsConfig {
 pub struct TestDefinition {
     pub id: Option<Thing>,
     pub name: String,
-    pub workers: u32,
-    pub iterations: u32,
+    pub workers: u8,
+    pub iterations: u8,
     pub dataset_name: String,
     pub algo_list: Vec<String>,
     #[serde(rename = "test_type")]
